services/accelerate-agentic-service/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0,<1.0.0
pydantic-settings>=2.0.0,<3.0.0
httpx>=0.28.0
pybreaker>=1.4.0,<2.0.0
prometheus-client>=0.22.0
langgraph>=1.1.0
langgraph-checkpoint-postgres>=3.0.0
psycopg[binary]>=3.2.0
langchain-core>=1.2.18
langchain-anthropic>=1.3.4
langchain-openai>=1.1.11
langchain-google-genai>=4.2.1
ag-ui-protocol>=0.1.13
langchain>=1.0.0
langchain-mcp-adapters>=0.2.2
pyyaml>=6.0
langfuse[langchain]>=4.0.0
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0,<5.0.0
respx>=0.21.0
setuptools>=78.1.1
sentry-sdk==2.54.0
gcloud-aio-storage>=9.0.0
google-genai>=1.0.0
Pillow>=10.0.0
crawl4ai>=0.8.5
huggingface-hub>=0.33.2,<1.0.0
playwright>=1.58.0
instructor>=1.14.0
